put getNetwork in DiagramUtils
