+++ src/main/java/com/powsybl/sld/server/SingleLineDiagramService.java
@@ -9,7 +9,6 @@ package com.powsybl.sld.server;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.network.store.client.NetworkStoreService;
-import com.powsybl.network.store.client.PreloadingStrategy;
 import com.powsybl.sld.SingleLineDiagram;
 import com.powsybl.sld.layout.*;
 import com.powsybl.sld.library.ComponentLibrary;
@@ -47,22 +46,6 @@ class SingleLineDiagramService {
     @Autowired
     private NetworkStoreService networkStoreService;
 
-    public static Network getNetwork(UUID networkUuid, String variantId, NetworkStoreService networkStoreService, PreloadingStrategy preloadingStrategy) {
-        try {
-            Network network = networkStoreService.getNetwork(networkUuid, preloadingStrategy);
-            if (variantId != null) {
-                network.getVariantManager().setWorkingVariant(variantId);
-            }
-            return network;
-        } catch (PowsyblException e) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
-        }
-    }
-
-    public static Network getNetwork(UUID networkUuid, String variantId, NetworkStoreService networkStoreService) {
-        return getNetwork(networkUuid, variantId, networkStoreService, null);
-    }
-
     private static SubstationLayoutFactory getSubstationLayoutFactory(String substationLayout) {
         SubstationLayoutFactory substationLayoutFactory;
         switch (substationLayout) {
@@ -80,7 +63,7 @@ class SingleLineDiagramService {
     }
 
     Pair<String, String> generateSvgAndMetadata(UUID networkUuid, String variantId, String id, SingleLineDiagramParameters diagParams) {
-        Network network = getNetwork(networkUuid, variantId, networkStoreService);
+        Network network = DiagramUtils.getNetwork(networkUuid, variantId, networkStoreService);
         if (network.getVoltageLevel(id) == null && network.getSubstation(id) == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Voltage level or substation " + id + " not found");
         }
